+++ Source/MMElement.h
@@ -25,6 +25,7 @@
 
 #import <Foundation/Foundation.h>
 
+@class MMElement;
 
 typedef enum
 {
@@ -76,8 +77,8 @@ typedef NS_ENUM(NSInteger, MMTableCellAlignment)
 @property (copy,   nonatomic) NSString      *identifier;
 @property (copy,   nonatomic) NSString      *stringValue;
 
-@property (assign, nonatomic) MMElement *parent;
-@property (copy,   nonatomic) NSArray   *children;
+@property (weak, nonatomic) MMElement *parent;
+@property (copy,   nonatomic) NSArray<MMElement *>   *children;
 
 @property (copy,   nonatomic) NSString  *language;