+++ Source/MMMarkdown.h
@@ -25,6 +25,11 @@
 
 #import <Foundation/Foundation.h>
 
+//! Project version number for MMMarkdown.
+FOUNDATION_EXPORT double MMMarkdownVersionNumber;
+
+//! Project version string for MMMarkdown.
+FOUNDATION_EXPORT const unsigned char MMMarkdownVersionString[];
 
 typedef NS_OPTIONS(NSUInteger, MMMarkdownExtensions)
 {
@@ -44,6 +49,7 @@ typedef NS_OPTIONS(NSUInteger, MMMarkdownExtensions)
     MMMarkdownExtensionsGitHubFlavored = MMMarkdownExtensionsAutolinkedURLs|MMMarkdownExtensionsFencedCodeBlocks|MMMarkdownExtensionsHardNewlines|MMMarkdownExtensionsStrikethroughs|MMMarkdownExtensionsTables|MMMarkdownExtensionsUnderscoresInWords,
 };
 
+NS_ASSUME_NONNULL_BEGIN
 @interface MMMarkdown : NSObject
 
 /*!
@@ -56,7 +62,7 @@ typedef NS_OPTIONS(NSUInteger, MMMarkdownExtensions)
  @result
     Returns an HTML string.
  */
-+ (NSString *)HTMLStringWithMarkdown:(NSString *)string error:(__autoreleasing NSError **)error __attribute__((nonnull(1)));
++ (NSString *)HTMLStringWithMarkdown:(NSString *)string error:(NSError * __autoreleasing * _Nullable)error;
 
 /*!
  Convert a Markdown string to HTML.
@@ -70,6 +76,7 @@ typedef NS_OPTIONS(NSUInteger, MMMarkdownExtensions)
  @result
     Returns an HTML string.
  */
-+ (NSString *)HTMLStringWithMarkdown:(NSString *)string extensions:(MMMarkdownExtensions)extensions error:(__autoreleasing NSError **)error __attribute__((nonnull(1)));
++ (NSString *)HTMLStringWithMarkdown:(NSString *)string extensions:(MMMarkdownExtensions)extensions error:(NSError * __autoreleasing * _Nullable)error;
 
 @end
+NS_ASSUME_NONNULL_END