+++ Source/MMParser.h
@@ -30,10 +30,12 @@
 
 @class MMDocument;
 
+NS_ASSUME_NONNULL_BEGIN
 @interface MMParser : NSObject
 
 - (id)initWithExtensions:(MMMarkdownExtensions)extensions;
 
-- (MMDocument *)parseMarkdown:(NSString *)markdown error:(__autoreleasing NSError **)error;
+- (MMDocument *)parseMarkdown:(NSString *)markdown error:(NSError * __autoreleasing * _Nullable)error;
 
 @end
+NS_ASSUME_NONNULL_END