+++ Source/MMSpanParser.m
@@ -41,6 +41,7 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
 
 @property (strong, nonatomic) MMElement *blockElement;
 @property (assign, nonatomic) BOOL parseEm;
+@property (assign, nonatomic) BOOL parseImages;
 @property (assign, nonatomic) BOOL parseLinks;
 @property (assign, nonatomic) BOOL parseStrong;
 @end
@@ -58,6 +59,7 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
         _extensions = extensions;
         _htmlParser = [MMHTMLParser new];
         self.parseEm     = YES;
+        self.parseImages = YES;
         self.parseLinks  = YES;
         self.parseStrong = YES;
     }
@@ -272,13 +274,16 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
             return element;
         
         [scanner beginTransaction];
-        element = [self _parseImageWithScanner:scanner];
+        element = [self _parseLinkWithScanner:scanner];
         [scanner commitTransaction:element != nil];
         if (element)
             return element;
-        
+    }
+    
+    if (self.parseImages)
+    {
         [scanner beginTransaction];
-        element = [self _parseLinkWithScanner:scanner];
+        element = [self _parseImageWithScanner:scanner];
         [scanner commitTransaction:element != nil];
         if (element)
             return element;
@@ -338,7 +343,7 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
 {
     NSCharacterSet        *alphanumerics = NSCharacterSet.alphanumericCharacterSet;
     NSMutableCharacterSet *boringChars = [alphanumerics mutableCopy];
-    [boringChars addCharactersInString:@",_-/:?&;%~!#+"];
+    [boringChars addCharactersInString:@",_-/:?&;%~!#+=@"];
     
     NSUInteger parenLevel = 0;
     while (1)
@@ -529,8 +534,13 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
 
 - (MMElement *)_parseEmAndStrongWithScanner:(MMScanner *)scanner
 {
+    // Must have 1-3 *s or _s
+    unichar character = scanner.nextCharacter;
+    if (!(character == '*' || character == '_'))
+        return nil;
+    
     NSCharacterSet *alphanumericSet = NSCharacterSet.alphanumericCharacterSet;
-    if (self.extensions & MMMarkdownExtensionsUnderscoresInWords)
+    if (self.extensions & MMMarkdownExtensionsUnderscoresInWords && character == '_')
     {
         // GFM doesn't italicize parts of words
         
@@ -543,11 +553,6 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
         if (isWordChar)
             return nil;
     }
-    
-    // Must have 1-3 *s or _s
-    unichar character = scanner.nextCharacter;
-    if (!(character == '*' || character == '_'))
-        return nil;
     
     // Must not be preceded by one of the same
     if (scanner.previousCharacter == character)
@@ -591,7 +596,7 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
         if (numberOfEndChars == 0 || (numberOfEndChars != remainingChars && remainingChars != 3))
             return NO;
         
-        if (self.extensions & MMMarkdownExtensionsUnderscoresInWords)
+        if (self.extensions & MMMarkdownExtensionsUnderscoresInWords && character == '_')
         {
             // GFM doesn't italicize parts of words
             unichar nextChar = scanner.nextCharacter;
@@ -815,9 +820,12 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     NSString *linkText  = [scanner.string substringWithRange:linkRange];
     
     // Make sure it looks like a link
-    NSRegularExpression *regex;
+    static NSRegularExpression *regex;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        regex = [NSRegularExpression regularExpressionWithPattern:@"^(\\w+)://" options:0 error:nil];
+    });
     NSRange matchRange;
-    regex      = [NSRegularExpression regularExpressionWithPattern:@"^(\\w+)://" options:0 error:nil];
     matchRange = [regex rangeOfFirstMatchInString:linkText options:0 range:NSMakeRange(0, linkText.length)];
     if (matchRange.location == NSNotFound)
         return nil;
@@ -892,11 +900,14 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     NSString *linkText  = [scanner.string substringWithRange:linkRange];
     
     // Make sure it looks like a link
-    NSRegularExpression *regex;
+    static NSRegularExpression *regex;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        regex = [NSRegularExpression regularExpressionWithPattern:@"^[-._0-9\\p{L}]+@[-\\p{L}0-9][-.\\p{L}0-9]*\\.\\p{L}+$"
+                                                          options:NSRegularExpressionCaseInsensitive
+                                                            error:nil];
+    });
     NSRange matchRange;
-    regex      = [NSRegularExpression regularExpressionWithPattern:@"^[-._0-9\\p{L}]+@[-\\p{L}0-9][-.\\p{L}0-9]*\\.\\p{L}+$"
-                                                           options:NSRegularExpressionCaseInsensitive
-                                                             error:nil];
     matchRange = [regex rangeOfFirstMatchInString:linkText options:0 range:NSMakeRange(0, linkText.length)];
     if (matchRange.location == NSNotFound)
         return nil;
@@ -982,6 +993,7 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     if (scanner.nextCharacter != '(')
         return nil;
     [scanner advance];
+    [scanner skipWhitespace];
     
     NSUInteger      urlLocation = scanner.location;
     NSUInteger      urlEnd      = urlLocation;
@@ -1012,6 +1024,13 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
         {
             if (level != 1)
                 return nil;
+            
+            [scanner skipWhitespace];
+            if (scanner.nextCharacter == ')')
+            {
+                [scanner advance];
+                level -= 1;
+            }
             break;
         }
         urlEnd = scanner.location;
@@ -1024,9 +1043,6 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     // If the level is still 1, then we hit a space.
     if (level == 1)
     {
-        // skip the whitespace
-        [scanner skipCharactersFromSet:NSCharacterSet.whitespaceCharacterSet];
-        
         // make sure there's a "
         if (scanner.nextCharacter != '"')
             return nil;
@@ -1091,6 +1107,8 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     
     // Look for the second []
     NSArray *idRanges = [self _parseLinkTextBodyWithScanner:scanner];
+    if (!idRanges)
+        return nil;
     if (!idRanges.count)
     {
         idRanges = element.innerRanges;
@@ -1149,7 +1167,9 @@ static NSString * const ESCAPABLE_CHARS = @"\\`*_{}[]()#+-.!>";
     // Add a transaction to protect the ! that was scanned
     [scanner beginTransaction];
     
+    self.parseImages = NO;
     element = [self _parseInlineLinkWithScanner:scanner];
+    self.parseImages = YES;
     
     if (element == nil)
     {