+++ Tests/MMLinkTests.m
@@ -174,6 +174,24 @@
     MMAssertMarkdownEqualsHTML(@"An empty [ by itself", @"<p>An empty [ by itself</p>");
 }
 
+- (void)testInlineLinkWithImage
+{
+    MMAssertMarkdownEqualsHTML(
+        @"[![Image](http://example.com/image.jpg)](http://www.example.com)",
+        @"<p><a href=\"http://www.example.com\"><img src=\"http://example.com/image.jpg\" alt=\"Image\" /></a></p>"
+    );
+}
+
+- (void)testInlineLinkWithPadding
+{
+    MMAssertMarkdownEqualsHTML(@"[test]( http://www.test.com )", @"<p><a href=\"http://www.test.com\">test</a></p>");
+}
+
+- (void)testInlineLinkInsideBrackets
+{
+    MMAssertMarkdownEqualsHTML(@"[[test](http://www.test.com)]", @"<p>[<a href=\"http://www.test.com\">test</a>]</p>");
+}
+
 
 #pragma mark - Reference Link Tests
 
@@ -304,6 +322,15 @@
 - (void)testReferenceLinkWithNoReference
 {
     MMAssertMarkdownEqualsHTML(@"[Foo][bar]", @"<p>[Foo][bar]</p>");
+}
+
+- (void)testLinkRecognitionWithAtSymbol
+{
+    // Use a string comparison for this test because the output is not valid XML
+    NSString *markdown  = @"Have you seen this medium link? https://medium.com/@philipla/the-two-types-of-product-virality-8ae744b1c4d7 It's Great";
+    NSString *generated = [MMMarkdown HTMLStringWithMarkdown:markdown extensions:MMMarkdownExtensionsAutolinkedURLs error:nil];
+    NSString *expected  =  @"<p>Have you seen this medium link? <a href=\"https://medium.com/@philipla/the-two-types-of-product-virality-8ae744b1c4d7\">https://medium.com/@philipla/the-two-types-of-product-virality-8ae744b1c4d7</a> It's Great</p>\n";
+    XCTAssertEqualObjects(generated, expected, @"HTML didn't match expected value");
 }